+++ src/boomerang-cli/CommandlineDriver.cpp
@@ -281,22 +281,6 @@ int CommandlineDriver::applyCommandline(const QStringList &args)
             m_project->getSettings()->useTypeAnalysis = false;
             continue;
         }
-        else if (arg == "-p") {
-            if (++i == args.size()) {
-                help();
-                return 1;
-            }
-
-            bool converted                           = false;
-            m_project->getSettings()->numToPropagate = args[i].toInt(&converted, 0);
-            if (!converted) {
-                std::cerr << "'-p': Bad argument '" << args[i].toStdString() << "' (try --help)."
-                          << std::endl;
-                return 1;
-            }
-
-            continue;
-        }
         else if (arg == "-sf") {
             if (++i == args.size()) {
                 help();