+++ src/boomerang/core/Settings.h
@@ -64,7 +64,6 @@ public:
 
     /// The file in which the dotty graph is saved
     QString dotFile;
-    int numToPropagate  = -1;
     bool usePromotion   = true;
     bool debugGen       = false;
     bool nameParameters = true;