+++ src/boomerang/ssl/statements/Statement.h
@@ -270,7 +270,7 @@ public:
 
     /// Experimental: may want to propagate flags first,
     /// without tests about complexity or the propagation limiting heuristic
-    bool propagateFlagsTo(Settings *settings);
+    bool propagateFlagsTo();
 
     /// simpify internal expressions
     /// \sa ExpSimplifier
@@ -311,8 +311,7 @@ public:
     /// Propagate to e from definition statement def.
     /// \returns true if a change made
     /// \note this procedure does not control what part of this statement is propagated to
-    bool doPropagateTo(const SharedExp &e, const std::shared_ptr<Assignment> &def,
-                       Settings *settings);
+    bool doPropagateTo(const SharedExp &e, const std::shared_ptr<Assignment> &def);
 
 private:
     /// replace a use of def->getLeft() by def->getRight() in this statement